ggml-hexagon: release ggml-dsp v0.60 on cDSP side
