+++ scripts/ggml-hexagon.cfg
@@ -1,5 +1,9 @@
 [general]
+#version of ggml-hexagon.cpp on ARM-AP side
 version = "1.00"
+#version of ggml-dsp.c on cDSP side
+ggmldsp_version = "0.60"
+
 #0: HEXAGON_BACKEND_QNNCPU
 #1: HEXAGON_BACKEND_QNNGPU
 #2: HEXAGON_BACKEND_QNNNPU / HEXAGON_BACKEND_CDSP
